Fixed Swagger UI
Using @ns.marshal_with(appointment_details) here seems to break Swagger UI
because the nested field dictionaries (yet used as per the documentation
https://flask-restplus.readthedocs.io/en/0.12.1/marshalling.html#complex-structures)
are not strictly model objects. No amounts of fields.Raw({ ... }) and fields.Nested({ ... }) healed it.

I'm unwilling to fight with the framework on that point,
so I'll just use the marshal() function and sacrifice the Swagger documentation for this particular endpoint,
which I think is a better compromise that giving up on the entire Swagger UI.